Fix initial rendering bugs (#1)
* Update lock file

* Fix Next.js hydration error

On page load, Next.js errors and links to [documentation][1] around a hydration error. The console has this warning:

```
Warning: In HTML, <th> cannot be a child of <thead>.
```

Wrapping in `<tr />` addresses this and conforms to [documentation on `<thead />`][2]

[1]: https://nextjs.org/docs/messages/react-hydration-error
[2]: https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/thead

* Ensure each child in list has a `key` prop

Keys are important for more performant updates when React re-renders a list. See [React documentation].

[1]: https://reactjs.org/link/warning-keys

* Use compound key to avoid collisions